added plans and orders schema along with updation

+++ controllers/owner.controller.js
@@ -1,4 +1,4 @@
-const { validateMessProfile, fieldValidation_emailVerification }= require('../validators/owner.validation')
+const { validateMessProfile, fieldValidation_emailVerification, validateMessPlan }= require('../validators/owner.validation')
 const MessProfile= require('../models/mess.schema')
 const { sequelize } = require('../services/connection_services_')
 const { uploadFileToS3 }= require('../services/s3FileUpload_services')
@@ -87,7 +87,7 @@ exports.createMessProfile = async (req, res) => {
       await t.commit()
       
       console.log(`otp sent to: ${email}`)
-      return res.status(201).json({ success: true, message: 'Mess profile created.Email verification needed', email: email, context: otpRecord.context, requestId: otpRecord.requestId, messId: mess.messId })
+      return res.status(201).json({ success: true, message: 'Mess profile created.Email verification needed', otp: otpRecord.otp, email: email, context: otpRecord.context, requestId: otpRecord.requestId, messId: mess.messId })
   
     } catch (err) {
       await t.rollback()
@@ -138,12 +138,33 @@ exports.handlePostVerifyMessEmail = async (req, res) => {
       messProfile.isEmailVerified = true
       messProfile.status= 'activated'
       await messProfile.save({ transaction: t })
+
+      const ownerProfile = await OwnerProfile.findOne({
+        where: { userId: messProfile.messOwnerId },
+        transaction: t,
+        lock: true
+      })
+
+      if (!ownerProfile) {
+        await t.rollback();
+        return res.status(404).json({ success: false, message: 'Owner profile not found.' });
+      }
+      
+      if (!ownerProfile.mess_ids.includes(messId)) {
+        ownerProfile.set({
+          mess_ids: [...ownerProfile.mess_ids, messId],
+          messCount: ownerProfile.messCount + 1
+        })
+        await ownerProfile.save({ transaction: t })
+      }
   
       await otpRecord.destroy({ transaction: t })
   
       await t.commit()
-  
+      console.log('Mess verified successfully.')
+
       return res.status(200).json({ success: true, message: 'Email verified successfully.' })
+
     } catch (err) {
       if (t) await t.rollback()
       console.error('Error verifying mess email:', err)
@@ -253,5 +274,72 @@ exports.updateMessProfile = async (req, res) => {
         if (t) await t.rollback();
         console.error('Update Mess Error:', err)
         return res.status(500).json({ success: false, message: 'Internal server error' })
+  }
+}
+
+
+exports.createMessPlan = async (req, res) => {
+  try {
+      if (!req.file) {
+        return res.status(400).json({ success: false, message: 'Image file is required.' })
+      }
+
+      let parsedMenu
+      try {
+        parsedMenu = JSON.parse(req.body.menu)
+      } catch (err) {
+        console.error('error in create plan', err.message)
+        return res.status(400).json({ success: false, message: 'Invalid JSON format for menu.' })
+      }
+
+      const { error, value } = validateMessPlan.validate({
+            messId: req.body.messId,
+            name: req.body.name,
+            description: req.body.description,
+            menu: parsedMenu,
+            durationDays: req.body.expiryDays,
+            price: req.body.price,
+      })
+
+      if (error) {
+        return res.status(400).json({ success: false, message: error.details[0].message })
+      }
+      
+      const userId = req.user?.id
+      const mess = await MessProfile.findOne({
+        where: {
+          messId: req.body.messId,
+          messOwnerId: userId
+        }
+      })
+  
+      if (!mess) {
+        return res.status(403).json({ success: false, message: 'You are not authorized to create plans for this mess.' })
+      }
+
+      const file = req.file
+      const aws_folder= `test`
+      const s3Url = await uploadFileToS3(file, aws_folder)
+
+      const currentDate = new Date()
+      const expiryDate = new Date(currentDate)
+      expiryDate.setDate(expiryDate.getDate() + parseInt(req.body.expiryDays))
+
+      const newPlan = await MessPlan.create({
+        messId: req.body.messId,
+        name: req.body.name,
+        description: req.body.description,
+        menu: parsedMenu,
+        durationDays: req.body.expiryDays,
+        expiryDate,
+        imageUrl: s3Url,
+        price: req.body.price
+      })
+
+      return res.status(201).json({ success: true, message: 'Mess plan created successfully', data: newPlan })
+
+  } catch (err) {
+      console.error('MessPlan Creation Error:', err)
+      return res.status(500).json({ success: false, message: 'Internal Server Error' })
   }
 }
